move parse method
