+++ internal/CMakeLists.txt
@@ -32,4 +32,5 @@ add_library(webserv_internal STATIC
         task/write_file.hpp
         http/context.hpp
         http/context.cpp
+        http/parse_raw_request.cpp
 )
