+++ internal/http/request.cpp
@@ -53,13 +53,3 @@ Option<std::string> Request::header(const std::string &key) const {
 const std::string &Request::text() const {
     return body_;
 }
-
-Request Request::parseRawRequest(const std::string &raw_request) {
-    // TODO: パースの処理をかく
-    return Request(
-            kMethodUnknown,
-            "/",
-            std::map<std::string, std::string>(),
-            std::map<std::string, std::string>(),
-            raw_request);
-}
